feature/refactor-easy-activity
- Refactor EasyActivity
